pyproject.toml
# Copyright 2025 Ubie, Inc.
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#      https://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

[project]
name = "lightdash-pre-commit-hooks"
version = "0.1.0"
description = "Lightdash pre-commit hooks"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Information Technology",
  "Intended Audience :: System Administrators",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Typing :: Typed",
]
dependencies = ["pydantic>=2.8,<3.0", "pyyaml>=6.0.2"]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.5",
  "bandit>=1.8.3",
  "isort>=6.0.1",
  "ruff>=0.9.10",
  "mypy>=1.15.0",
  "pylint>=3.3.5",
  "datamodel-code-generator>=0.30,<0.40",
  "pre-commit>=4.2.0",
  "types-PyYAML>=6.0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
exclude = [
  ".git",
  ".venv",
  ".mypy_cache",
  ".pytest_cache",
  ".github",
  ".trunk",
  "tests",
  "dev",
  "dist",
  ".python-version",
  "resources",
]

[tool.hatch.build.targets.wheel]
packages = ["src/lightdash_pre_commit"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = ["-v", "-s", "--tb=short"]
